wip: moved the exception handler to generator_utils
